Move trackReady function over to new file
