+++ app/frontend/js/thyme/chapter_manager.js
@@ -1,4 +1,5 @@
-import { onVideoMetadataLoaded, renderLatex } from "./utility";
+import { onTrackReady } from "./video_events";
+import { renderLatex } from "./utility";
 
 /**
   This file wraps up most functionality of the thyme player(s) concerning chapters.
@@ -15,42 +16,7 @@ export class ChapterManager {
    * It receives a boolean value that indicates whether chapters are present.
    */
   load(onLoad) {
-    const chapters = this.#getChapters();
-    const chapterManager = this;
-    let notified = false;
-
-    // "Chapters are ready" is tied to the chapters track element itself, not to a
-    // single video event. onVideoMetadataLoaded/canplay can fire before the track
-    // has parsed its cues (readyState 2) — notably when the video is already
-    // buffered (e.g. under test). The old code only called onLoad from the
-    // loadedmetadata handler, and only if the track happened to be ready at that
-    // instant; otherwise onLoad never fired, onVideoDataReady hung, and the
-    // content sidebar stayed hidden (flaky in CI). Listening to the track's own
-    // load/error events as well makes onLoad fire exactly once, whichever wins.
-    const notifyIfReady = () => {
-      if (notified) {
-        return;
-      }
-      if (chapters.readyState === 2) {
-        notified = true;
-        chapterManager.#displayChapters();
-        if (onLoad) {
-          onLoad(chapters.track ? (chapters.track.cues.length > 0) : false);
-        }
-      }
-      else if (chapters.readyState === 3) {
-        notified = true;
-        if (onLoad) {
-          onLoad(false);
-        }
-      }
-    };
-
-    notifyIfReady();
-    chapters.addEventListener("load", notifyIfReady);
-    chapters.addEventListener("error", notifyIfReady);
-    onVideoMetadataLoaded(thymeAttributes.video, notifyIfReady);
-    thymeAttributes.video.addEventListener("canplay", notifyIfReady);
+    onTrackReady(this.#getChapters(), () => this.#displayChapters(), onLoad);
   }
 
   previousChapterStart() {
